+++ common/src/main/java/com/chefmooon/colourfulclocks/integration/jei/JEIPlugin.java
@@ -1,0 +1,8 @@
+package com.chefmooon.colourfulclocks.integration.jei;
+
+import com.chefmooon.colourfulclocks.common.util.TextUtil;
+import net.minecraft.resources.ResourceLocation;
+
+public class JEIPlugin {
+    public static final ResourceLocation ID = TextUtil.res("jei_plugin");
+}